% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residencysearch.R
\name{aggdata2store}
\alias{aggdata2store}
\title{Aggregated data to data store}
\usage{
aggdata2store(dbConnection, aggData)
}
\arguments{
\item{dbConnection}{a DBI database connection}

\item{aggData}{a data frame containing aggregated data}
}
\value{
Nothing
}
\description{
Migrate the aggregated data to the data store
}
\details{
This functions writes the aggregated data created
with residencysearch() to the data store.
}
\examples{
\dontrun{aggdata2store(dbConnection, aggData)}
}

