% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input2store.R
\name{input2store}
\alias{input2store}
\title{Migrate input data to data store}
\usage{
input2store(directory)
}
\arguments{
\item{directory}{The directory containing the tracking data}
}
\value{
Nothing
}
\description{
Read all the tracking data in a directory, validate it and
write it to the data store.
}
\examples{
input2store("/path/to/input/directory/")
}

