% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input2store.R
\name{merge_files}
\alias{merge_files}
\title{Read all tracking data in a directory}
\usage{
merge_files(directory)
}
\arguments{
\item{directory}{The directory containing all tracking data input files}
}
\value{
All tracking data of the different input files as one data frame
with the columns Date.Time, Receiver.id, Station.Name, Transmitter.id
}
\description{
Merge all input files containing tracking data that are located in a single
directory
}
\examples{
\dontrun{merge_files("/path/to/directory/")}
}

