% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input2store.R
\name{parse_date}
\alias{parse_date}
\title{Parse a tracking data timestamp}
\usage{
parse_date(dateStr)
}
\arguments{
\item{dateStr}{A vector of strings containing date times.}
}
\value{
A vector of parsed date times. For each parsed input string, the return
value is a date time as character in the format "Y-m-d H:M:S" if the input
string has the format "Y-m-d H:M:S" or "d-m-Y H:M:S". If not, NA is returned.
}
\description{
Return a date time (as character) if the input character contains date time
data in a known format. Else, return NA.
}
\examples{
# should return c("2000-03-21 13:21:42", "1952-04-15 09:00:31", NA, NA)
parse_date(c("2000-03-21 13:21:42", "15-04-1952 09:00:31", "03-31-2004 03:49:23", "31-02-2004 04:29:42"))
}

