% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input2store.R
\name{parse_station}
\alias{parse_station}
\title{Parse stations names from input}
\usage{
parse_station(stations)
}
\arguments{
\item{stations}{Vector containing station names}
}
\value{
Vector containing valid station names or NA's
}
\description{
parse_station will check whether all fields in the input vector
are valid station names using a regular expression. A vector of the same
length is returned containing the station names for valid ones, and NA
for invalid ones.
}

