% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input2store.R
\name{read_input}
\alias{read_input}
\title{Read tracking data file}
\usage{
read_input(filename)
}
\arguments{
\item{filename}{The name of the input file to read.}
}
\value{
The tracking data from the input file as a data frame with the
following columns: Date.Time, Receiver.id, Receiver.Name, Transmitter.id
}
\description{
Read a file containing tracking data and convert it to a
standard data frame.
}
\examples{
read_input("VR2W_122340_20141010_1.csv")
}

