% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residencysearch.R
\name{residencysearch}
\alias{residencysearch}
\title{Residency Search}
\usage{
residencysearch(detections, maxAbsence = 59)
}
\arguments{
\item{detections}{The data frame containing raw detection data}

\item{maxAbsence}{The maximum number of minutes that a fish is
allowed to be absent from a receiver. In other words, if the next
detection exceeds this limit, a new time interval will be created.}
}
\value{
Aggregated data as a data frame containing the columns
"Transmitter.id", "Station.Name", "Start" and "End".
}
\description{
Aggregate input data by performing residency search
}
\details{
Raw detection data can contain a lot of detections
when a fish is not migrating and staying near a receiver.
These detections can be aggregated by calculating the time
interval during which a fish was located at a certain receiver.
Time intervals have a start time and end time. Creating these
time intervals based on the raw detections is called residency
search.
}
\examples{
\dontrun{
inputdata <- input2store(dbConnection, "/path/to/raw/data/")
residencysearch(inputdata)
}
}

