% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input2store.R
\name{validate_data}
\alias{validate_data}
\title{Validate input tracking data}
\usage{
validate_data(indata)
}
\arguments{
\item{indata}{A data frame containing tracking data}
}
\value{
TRUE if data is ok
}
\description{
Validates a given data frame containing tracking data and returns
TRUE if everything is ok or an error of something is wrong
}
\examples{
\dontrun{validate_data(data)}
}

